Chap02/UseModelServing/FlaskWebModelServing/requirements.txt
absl-py==0.9.0
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
Cython==0.29.7
DateTime==4.3
decorator==4.4.0
deepmachine==0.7.5
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.8.4
grpcio==1.17.1
h5py==2.9.0
idna==2.6
imageio==2.5.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.11.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.1.0
mistune==0.8.4
mock==3.0.5
mxnet==1.3.1
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numpy==1.18.1
oauthlib==3.1.0
onnx==1.6.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
Pillow==6.0.0
pip==20.0.2
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
Pygments==2.4.0
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.4.1
Send2Trash==1.5.0
setuptools==45.2.0
six==1.12.0
SpeechRecognition==3.8.1
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
tensorflow-gpu-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
typing-extensions==3.7.4.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.0
wheel==0.32.3
widgetsnbextension==3.4.2
wrapt==1.12.0
zope.interface==4.6.0
